openclaw-firetitan/requirements.txt
fireworks-ai[training]
httpx
fastapi
uvicorn
transformers
torch
